refactor: split peg alert settings view
